+++ pkgs/test_core/lib/test_core.dart
@@ -6,12 +6,8 @@
     'Please use package:test.')
 library test_core;
 
-export 'package:test_api/expect.dart';
 export 'package:test_api/hooks.dart' show TestFailure;
 // Not yet deprecated, but not exposed through focused libraries.
 export 'package:test_api/test_api.dart' show registerException;
-// Deprecated exports not surfaced through focused libraries.
-export 'package:test_api/test_api.dart'
-    show ErrorFormatter, expectAsync, throws, Throws;
 
 export 'scaffolding.dart';